requirements.txt
requests==2.31.0
langchain==0.1.4
openai==1.10.0
python-dotenv==1.0.1
serpapi==0.1.5
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
